Refactor transformer example
